refactor(cleanup): hoist parse_enabled_env into runtara-core
Replaces four copies of the same 8-line `parse_enabled_env` helper
(plus four copies of its table-driven unit test) with a single
`pub fn` in `runtara-core::config`, imported from each worker.

Both `runtara-server` and `runtara-environment` already depend on
`runtara-core`, so no Cargo.toml change is needed. The runtara-core
test reuses the existing `EnvGuard` + `ENV_MUTEX` pattern in
`config.rs`, which is more disciplined than the inline
`unsafe { std::env::set_var }` calls each worker had.

Behavior is unchanged: cleanup is on by default; only an explicit
opt-out value (`false`/`0`/`no`/`off`/`disabled`, case-insensitive,
trimmed) disables it.
