+++ crates/runtara-server/src/workers/invocation_cleanup_worker.rs
@@ -10,6 +10,7 @@
 use std::time::Duration;
 
 use chrono::Utc;
+use runtara_core::config::parse_enabled_env;
 use sqlx::PgPool;
 use tracing::{debug, error, info, warn};
 
@@ -246,20 +247,6 @@ impl InvocationCleanupWorker {
     }
 }
 
-/// Resolve a `*_CLEANUP_ENABLED` env var with a "default-on, opt-out only"
-/// rule. Cleanup is enabled unless the env var is explicitly set to a
-/// recognised false-like value (case-insensitive). Misconfigurations like
-/// `"yes"`, `"on"`, or a typo do **not** silently disable cleanup.
-fn parse_enabled_env(name: &str) -> bool {
-    match std::env::var(name) {
-        Ok(v) => !matches!(
-            v.trim().to_ascii_lowercase().as_str(),
-            "false" | "0" | "no" | "off" | "disabled"
-        ),
-        Err(_) => true,
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -290,32 +277,6 @@ mod tests {
         assert!(config.enabled, "Enabled-by-default expected");
         assert_eq!(config.max_age.as_secs() / 86400, 3);
         assert_eq!(config.metrics_max_age.as_secs() / 86400, 365);
-    }
-
-    #[test]
-    fn test_parse_enabled_env_default_on() {
-        const VAR: &str = "RUNTARA_TEST_INVOCATION_CLEANUP_ENABLED_PARSE";
-        unsafe { std::env::remove_var(VAR) };
-        assert!(parse_enabled_env(VAR), "unset must be enabled");
-
-        for v in [
-            "true",
-            "1",
-            "yes",
-            "on",
-            "True",
-            "TRUE",
-            "anything-else",
-            "",
-        ] {
-            unsafe { std::env::set_var(VAR, v) };
-            assert!(parse_enabled_env(VAR), "{v:?} must NOT silently disable");
-        }
-        for v in ["false", "0", "no", "off", "disabled", "FALSE", "Off"] {
-            unsafe { std::env::set_var(VAR, v) };
-            assert!(!parse_enabled_env(VAR), "{v:?} must explicitly disable");
-        }
-        unsafe { std::env::remove_var(VAR) };
     }
 
     #[tokio::test]
